Edge Bootstrapping and Splitting

Estimate the graph dimension via eigenvalue cross-validation (EigCV). A graph has dimension k if the first k eigenvectors of its adjacency matrix are correlated with its population eigenspace, and the others are not. Edge bootstrapping sub-samples the edges of the graph (without replacement). Edge splitting separates the edges into a training part and a testing part.

Usage

eigcv(
  A,
  k_max,
  bootstrap = 10,
  split = 0.1,
  alpha = 0.05,
  ptol = 1e-32,
  correct = "none",
  laplacian = TRUE,
  regularize = TRUE,
  align = FALSE,
  trace = FALSE
)

Arguments

A

the adjacency matrix, will be coerced to a Matrix class.

k_max

integer(1), number of eigenvectors to compute.

bootstrap

integer(1), number of graph bootstraps, default to 10. Graph bootstrapping is to account for the randomness in graph splitting, rather than obtaining any statistic (as a traditional bootstrap does). Hence, a small number (e.g., 3~10) of bootstraps usually suffices. If bootstrap>1, the test statistics will be averaged across bootstraps and the p-values will be calculated based on the averaged statistics.

split

numeric(1), the proportion of leave-out edges (into the test set). Must take value from (0,1).

alpha

numeric(1), significance level of each test, default to 0.05. This is used to cut off the dimension estimation.

ptol

numeric(1), the tolerance of minimal p-value.

correct

character(1), correction method for multiple testing. This is recommended if k_max is large. See p.adjust for the acceptable values of method. If correct is not "none", the returned summary table will have an extra column called padj that repeats the p-value in the previous column.

laplacian

logical(1), use the normalized and regularized adjacency matrix (i.e. L)

regularize

logical(1), return a regularized symmetric Laplacian. The default is TRUE. This is ignored if laplacian=FALSE.

align

logical(1), align the trained singular vectors with those of the full graph. The default is FALSE. This option is experimental and should be used with caution.

trace

logical(1), for diagnostic use. If TRUE, return additionally cv_stats containing the test statistics of every bootstrap, and if align=TRUE, then the eigs of the full graph.

Value

A eigcv object, which contains:

inference

inferred graph dimension.

summary

summary table of the tests.

bootstrap

number of bootstraps performed.

split

graph splitting probability used.

alpha

significance level of each test.

trace

only if trace=TRUE, all bootstrapped statistics.